+++ scripts/generateCss.js
@@ -4,36 +4,14 @@
 
 "use strict";
 
-const fs = require("fs"),
-  sass = require("node-sass");
+// This script is designed to be executed from the scripts folder
+
+var fs = require("fs"),
+    sass = require("node-sass");
 
 // Generate css directory if not already existing.
 if (!fs.existsSync("../css")) {
-  fs.mkdirSync("../css");
-}
-
-/**
- * Generates a css file out of the input file.
- * @param {string} inputFilePath - the path of the input file.
- * @param {string} outputFilePath - the path of the output file.
- * @param {Boolean} compressed - true if the output file should be compressed.
- */
-function processScssFile(inputFilePath, outputFilePath, compressed) {
-  var outputStyle = "nested";
-  if (compressed)
-    outputStyle = "compressed";
-
-  sass.render({
-    file: inputFilePath,
-    outputStyle: outputStyle
-  }, function (error, result) {
-    if (!error) {
-      fs.writeFileSync(outputFilePath, result.css);
-    } else {
-      console.log("An error occurred while generating the css files out of the scss files:");
-      console.log(error.message);
-    }
-  });
+    fs.mkdirSync("../css");
 }
 
 /**
@@ -43,15 +21,23 @@ function processScssFile(inputFilePath, outputFilePath, compressed) {
  * @param {string} outputDirectory - the path of the directory the css files are written to.
  */
 function generateCss(inputDirectory, outputDirectory) {
-  var scssFiles = fs.readdirSync(inputDirectory);
+    var inputFiles = fs.readdirSync(inputDirectory);
 
-  console.log("Processing scss files...");
-  for (var i = 0; i < scssFiles.length; i++) {
-    var inputPath = inputDirectory + "/" + scssFiles[i],
-      outputPathWithoutEnding = outputDirectory + "/" + scssFiles[i].substring(0, scssFiles[i].indexOf(".scss"));
-
-    processScssFile(inputPath, outputPathWithoutEnding + ".css", false);
-  }
+    console.log("Processing scss files...");
+    for (var i = 0; i < inputFiles.length; i++) {
+        (function (i) {
+            sass.render({
+                file: inputDirectory + "/" + inputFiles[i],
+                outputStyle: "nested"
+            }, function (error, result) {
+                if (!error)
+                    fs.writeFileSync(outputDirectory + "/" + inputFiles[i].slice(0, -4) + "css", result.css);
+                else
+                    console.log(error.message);
+            })
+        }(i));
+    }
+    console.log("Finished processing scss files.");
 }
 
 generateCss("../scss/", "../css/");